removed algorithm combination
major change:
removed the algorithm combinations - AUI + SD, and AUMA + SD
these algorithms do not make sense to implement because they assume that
you already know the average and max cpu utilization. also, it is
impossible to know how the new start and end times of the runtimes that
were shut down.