fixed algorithms
fixed auma, cr, and sd logical mistakes
also changed the order in which data was modified. for example, it used
to run vm-based algorithms before running moer-based algorithm, which
was wrong because we cannot predict if the sd or cr will modify that vm
before it is run,
and therefore the vm-based algorithms should not affect the actions of
the moer-based algorithms. so, the algRunner has been changed to run the
moer-based algorithms first, then the vm-based algorithms.